+++ test-suite/run_tests.sh
@@ -35,7 +35,7 @@ ALL_TEST_NAMES=(
     "TestUnhandledExceptionInstance"
     "TestStdIO"
     "TestModules"
-    "TestTracePoint"
+    "TestLogpoints"
     "TestRemoteConsole"
     "TestSourceFileMap"
     "TestEvaluatePrimitiveUnary"